renamed fields-node.ts to fields.ts
