+++ packages/data/src/types/abstract-query/fields/nested.ts
@@ -1,5 +1,5 @@
 import type { AbstractQueryModifiers } from '../modifiers.js';
-import type { AbstractQueryFieldNode } from './field-nodes.js';
+import type { AbstractQueryFieldNode } from './fields.js';
 import type {
 	AbstractQueryFieldNodeNestedRelationalMany,
 	AbstractQueryFieldNodeNestedRelationalOne,
